+++ example/lib/analytics/counter_analytics.dart
@@ -1,6 +1,6 @@
 import 'package:loglytics/loglytics.dart';
 
-class CounterAnalytics extends AnalyticsData {
+class CounterAnalytics extends Analytics {
   final String counterButton = 'counter_button';
   final String incrementCounter = 'increment_counter';
 }