+++ lib/loglytics.dart
@@ -2,10 +2,14 @@ library loglytics;
 
 // src/analytics
 export 'src/analytics/analytic.dart';
+export 'src/analytics/analytics.dart';
 export 'src/analytics/analytics_interface.dart';
 export 'src/analytics/analytics_service.dart';
 // src/crashlytics
 export 'src/crash_reports/crash_reports_interface.dart';
-// src/services
+// src/enums
+export 'src/enums/log_type.dart';
+// src/log
+export 'src/log/custom_log.dart';
+// src/loglytics
 export 'src/loglytics/loglytics.dart';
-export 'src/loglytics/analytics_data.dart';