+++ lib/src/analytics/analytics_service.dart
@@ -1,4 +1,4 @@
-import 'package:loglytics/src/loglytics/analytics_data.dart';
+import 'package:loglytics/src/analytics/analytics.dart';
 
 import '../../loglytics.dart';
 import '../loglytics/loglytics.dart';
@@ -8,10 +8,10 @@ import 'analytics_interface.dart';
 /// Used to provide an easy interface for sending analytics.
 ///
 /// Each [AnalyticType] has its own method that exposes implementations of predefined
-/// [AnalyticsData].
+/// [Analytics].
 /// example you can use [AnalyticsService.tap] and it the will automatically provide you with the
-/// proper [AnalyticsData].
-class AnalyticsService<A extends AnalyticsData> {
+/// proper [Analytics].
+class AnalyticsService<A extends Analytics> {
   AnalyticsService({
     required A analyticsData,
     Loglytics? loglytics,
@@ -53,14 +53,14 @@ class AnalyticsService<A extends AnalyticsData> {
     _loglytics?.logAnalytic(name: name, value: _value);
   }
 
-  /// Sends a custom analytic event by providing both [AnalyticsData].
+  /// Sends a custom analytic event by providing both [Analytics].
   ///
   /// This method may be used to log anything that is not covered by any other method in this class
   /// and expects an [Analytic] in return from the [analytic] callback.
   void event({required Analytic Function(A analytics) analytic}) =>
       _logEvent(analytic(_analyticsData));
 
-  /// Sends an [AnalyticType.tap] and provides the appropriate [AnalyticsData].
+  /// Sends an [AnalyticType.tap] and provides the appropriate [Analytics].
   void tap({
     required String Function(A analytics) subject,
     Map<String, Object?>? Function(A analytics)? parameters,
@@ -73,7 +73,7 @@ class AnalyticsService<A extends AnalyticsData> {
         ),
       );
 
-  /// Sends an [AnalyticType.click] and provides the appropriate [AnalyticsData].
+  /// Sends an [AnalyticType.click] and provides the appropriate [Analytics].
   void click({
     required String Function(A analytics) subject,
     Map<String, Object?>? Function(A analytics)? parameters,
@@ -86,7 +86,7 @@ class AnalyticsService<A extends AnalyticsData> {
         ),
       );
 
-  /// Sends an [AnalyticType.focus] and provides the appropriate [AnalyticsData].
+  /// Sends an [AnalyticType.focus] and provides the appropriate [Analytics].
   void focus({
     required String Function(A analytics) subject,
     Map<String, Object?>? Function(A analytics)? parameters,
@@ -99,7 +99,7 @@ class AnalyticsService<A extends AnalyticsData> {
         ),
       );
 
-  /// Sends an [AnalyticType.select] and provides the appropriate [AnalyticsData].
+  /// Sends an [AnalyticType.select] and provides the appropriate [Analytics].
   void select({
     required String Function(A analytics) subject,
     Map<String, Object?>? Function(A analytics)? parameters,
@@ -112,7 +112,7 @@ class AnalyticsService<A extends AnalyticsData> {
         ),
       );
 
-  /// Sends an [AnalyticType.connect] and provides the appropriate [AnalyticsData].
+  /// Sends an [AnalyticType.connect] and provides the appropriate [Analytics].
   void connect({
     required String Function(A analytics) subject,
     Map<String, Object?>? Function(A analytics)? parameters,
@@ -125,7 +125,7 @@ class AnalyticsService<A extends AnalyticsData> {
         ),
       );
 
-  /// Sends an [AnalyticType.connect] and provides the appropriate [AnalyticsData].
+  /// Sends an [AnalyticType.connect] and provides the appropriate [Analytics].
   void disconnect({
     required String Function(A analytics) subject,
     Map<String, Object?>? Function(A analytics)? parameters,
@@ -138,7 +138,7 @@ class AnalyticsService<A extends AnalyticsData> {
         ),
       );
 
-  /// Sends an [AnalyticType.view] and provides the appropriate [AnalyticsData].
+  /// Sends an [AnalyticType.view] and provides the appropriate [Analytics].
   void view({
     required String Function(A analytics) subject,
     Map<String, Object?>? Function(A analytics)? parameters,
@@ -151,7 +151,7 @@ class AnalyticsService<A extends AnalyticsData> {
         ),
       );
 
-  /// Sends an [AnalyticType.hide] and provides the appropriate [AnalyticsData].
+  /// Sends an [AnalyticType.hide] and provides the appropriate [Analytics].
   void hide({
     required String Function(A analytics) subject,
     Map<String, Object?>? Function(A analytics)? parameters,
@@ -164,7 +164,7 @@ class AnalyticsService<A extends AnalyticsData> {
         ),
       );
 
-  /// Sends an [AnalyticType.open] and provides the appropriate [AnalyticsData].
+  /// Sends an [AnalyticType.open] and provides the appropriate [Analytics].
   void open({
     required String Function(A analytics) subject,
     Map<String, Object?>? Function(A analytics)? parameters,
@@ -177,7 +177,7 @@ class AnalyticsService<A extends AnalyticsData> {
         ),
       );
 
-  /// Sends an [AnalyticType.close] and provides the appropriate [AnalyticsData].
+  /// Sends an [AnalyticType.close] and provides the appropriate [Analytics].
   void close({
     required String Function(A analytics) subject,
     Map<String, Object?>? Function(A analytics)? parameters,
@@ -190,7 +190,7 @@ class AnalyticsService<A extends AnalyticsData> {
         ),
       );
 
-  /// Sends an [AnalyticType.fail] and provides the appropriate [AnalyticsData].
+  /// Sends an [AnalyticType.fail] and provides the appropriate [Analytics].
   void fail({
     required String Function(A analytics) subject,
     Map<String, Object?>? Function(A analytics)? parameters,
@@ -203,7 +203,7 @@ class AnalyticsService<A extends AnalyticsData> {
         ),
       );
 
-  /// Sends an [AnalyticType.success] and provides the appropriate [AnalyticsData].
+  /// Sends an [AnalyticType.success] and provides the appropriate [Analytics].
   void success({
     required String Function(A analytics) subject,
     Map<String, Object?>? Function(A analytics)? parameters,
@@ -216,7 +216,7 @@ class AnalyticsService<A extends AnalyticsData> {
         ),
       );
 
-  /// Sends an [AnalyticType.send] and provides the appropriate [AnalyticsData].
+  /// Sends an [AnalyticType.send] and provides the appropriate [Analytics].
   void send({
     required String Function(A analytics) subject,
     Map<String, Object?>? Function(A analytics)? parameters,
@@ -229,7 +229,7 @@ class AnalyticsService<A extends AnalyticsData> {
         ),
       );
 
-  /// Sends an [AnalyticType.receive] and provides the appropriate [AnalyticsData].
+  /// Sends an [AnalyticType.receive] and provides the appropriate [Analytics].
   void receive({
     required String Function(A analytics) subject,
     Map<String, Object?>? Function(A analytics)? parameters,
@@ -242,7 +242,7 @@ class AnalyticsService<A extends AnalyticsData> {
         ),
       );
 
-  /// Sends an [AnalyticType.valid] and provides the appropriate [AnalyticsData].
+  /// Sends an [AnalyticType.valid] and provides the appropriate [Analytics].
   void valid({
     required String Function(A analytics) subject,
     Map<String, Object?>? Function(A analytics)? parameters,
@@ -255,7 +255,7 @@ class AnalyticsService<A extends AnalyticsData> {
         ),
       );
 
-  /// Sends an [AnalyticType.invalid] and provides the appropriate [AnalyticsData].
+  /// Sends an [AnalyticType.invalid] and provides the appropriate [Analytics].
   void invalid({
     required String Function(A analytics) subject,
     Map<String, Object?>? Function(A analytics)? parameters,
@@ -268,7 +268,7 @@ class AnalyticsService<A extends AnalyticsData> {
         ),
       );
 
-  /// Sends an [AnalyticType.search] and provides the appropriate [AnalyticsData].
+  /// Sends an [AnalyticType.search] and provides the appropriate [Analytics].
   void search({
     required String Function(A analytics) subject,
     Map<String, Object?>? Function(A analytics)? parameters,
@@ -281,7 +281,7 @@ class AnalyticsService<A extends AnalyticsData> {
         ),
       );
 
-  /// Sends an [AnalyticType.like] and provides the appropriate [AnalyticsData].
+  /// Sends an [AnalyticType.like] and provides the appropriate [Analytics].
   void like({
     required String Function(A analytics) subject,
     Map<String, Object?>? Function(A analytics)? parameters,
@@ -294,7 +294,7 @@ class AnalyticsService<A extends AnalyticsData> {
         ),
       );
 
-  /// Sends an [AnalyticType.share] and provides the appropriate [AnalyticsData].
+  /// Sends an [AnalyticType.share] and provides the appropriate [Analytics].
   void share({
     required String Function(A analytics) subject,
     Map<String, Object?>? Function(A analytics)? parameters,
@@ -307,7 +307,7 @@ class AnalyticsService<A extends AnalyticsData> {
         ),
       );
 
-  /// Sends an [AnalyticType.comment] and provides the appropriate [AnalyticsData].
+  /// Sends an [AnalyticType.comment] and provides the appropriate [Analytics].
   void comment({
     required String Function(A analytics) subject,
     Map<String, Object?>? Function(A analytics)? parameters,
@@ -320,7 +320,7 @@ class AnalyticsService<A extends AnalyticsData> {
         ),
       );
 
-  /// Sends an [AnalyticType.input] and provides the appropriate [AnalyticsData].
+  /// Sends an [AnalyticType.input] and provides the appropriate [Analytics].
   ///
   /// Defaults to only sending the first analytic by settings [onlyFirstValue] to true.
   void input({
@@ -339,7 +339,7 @@ class AnalyticsService<A extends AnalyticsData> {
     _firstInput = analytic;
   }
 
-  /// Sends an [AnalyticType.increment] and provides the appropriate [AnalyticsData].
+  /// Sends an [AnalyticType.increment] and provides the appropriate [Analytics].
   void increment({
     required String Function(A analytics) subject,
     Map<String, Object?>? Function(A analytics)? parameters,
@@ -352,7 +352,7 @@ class AnalyticsService<A extends AnalyticsData> {
         ),
       );
 
-  /// Sends an [AnalyticType.decrement] and provides the appropriate [AnalyticsData].
+  /// Sends an [AnalyticType.decrement] and provides the appropriate [Analytics].
   void decrement({
     required String Function(A analytics) subject,
     Map<String, Object?>? Function(A analytics)? parameters,
@@ -365,7 +365,7 @@ class AnalyticsService<A extends AnalyticsData> {
         ),
       );
 
-  /// Sends an [AnalyticType.accept] and provides the appropriate [AnalyticsData].
+  /// Sends an [AnalyticType.accept] and provides the appropriate [Analytics].
   void accept({
     required String Function(A analytics) subject,
     Map<String, Object?>? Function(A analytics)? parameters,
@@ -378,7 +378,7 @@ class AnalyticsService<A extends AnalyticsData> {
         ),
       );
 
-  /// Sends an [AnalyticType.decline] and provides the appropriate [AnalyticsData].
+  /// Sends an [AnalyticType.decline] and provides the appropriate [Analytics].
   void decline({
     required String Function(A analytics) subject,
     Map<String, Object?>? Function(A analytics)? parameters,
@@ -391,7 +391,46 @@ class AnalyticsService<A extends AnalyticsData> {
         ),
       );
 
-  /// Sends the current screen and provides the appropriate [AnalyticsData].
+  /// Sends an [AnalyticType.scroll] and provides the appropriate [Analytics].
+  void scroll({
+    required String Function(A analytics) subject,
+    Map<String, Object?>? Function(A analytics)? parameters,
+  }) =>
+      _logEvent(
+        Analytic(
+          subject: subject(_analyticsData),
+          parameters: parameters?.call(_analyticsData),
+          type: AnalyticType.scroll,
+        ),
+      );
+
+  /// Sends an [AnalyticType.start] and provides the appropriate [Analytics].
+  void start({
+    required String Function(A analytics) subject,
+    Map<String, Object?>? Function(A analytics)? parameters,
+  }) =>
+      _logEvent(
+        Analytic(
+          subject: subject(_analyticsData),
+          parameters: parameters?.call(_analyticsData),
+          type: AnalyticType.start,
+        ),
+      );
+
+  /// Sends an [AnalyticType.stop] and provides the appropriate [Analytics].
+  void stop({
+    required String Function(A analytics) subject,
+    Map<String, Object?>? Function(A analytics)? parameters,
+  }) =>
+      _logEvent(
+        Analytic(
+          subject: subject(_analyticsData),
+          parameters: parameters?.call(_analyticsData),
+          type: AnalyticType.stop,
+        ),
+      );
+
+  /// Sends the current screen and provides the appropriate [Analytics].
   void screen({
     required String Function(A analytics) subject,
   }) {